W.I.P. 12-06-2024 10:16
added possibility to edit user and posts: This is grammatically correct.
as owner and own posts: This is clear. "Own posts" clarifies that editing applies to posts created by the user themselves (writer).
as writer of the post: This reinforces the previous point.
added little security to that only authorized people can do this: This sentence is a bit informal. You could consider "implemented security measures to ensure only authorized users can perform these actions."